ci(format-suggestions-on-pr): remove option to specify GitHub token
This option is not strictly necessary. For most typical repositories,
using the `GITHUB_TOKEN`/`github.token` (which the action uses by
default) will suffice. In the outlier cases, where a different token is
required to be explicitly specified, this can be achieved by specifying
an env variable named `REVIEWDOG_GITHUB_API_TOKEN` anyways.