ci: remove concurrency definitions from both reusable test workflows
Remove the concurrency definitions from both the reusable test workflows.
Test workflows are very likely to need to use a job matrix.
Given where they are defined (in the job hierarchy), it would become
unwieldy to manage the concurrency group definition, given all the
different possible options that can be configured in a job matrix.
This adds on unnecessary maintenance burden, since now you will have
to update the generic reusable workflows to cater to and specifically
address unique requirements, all to ensure that a new configuration in
the job matrix still runs and the respective (matrix) jobs spawned in
that workflow run don't get cancelled.