ci(formatting): always fail `format-check` job if formatting check fails
Always fail the `format-check` job if the formatting is not compliant
with the SciML style. This will ensure that the format checking is also
semantically correct and ensures that the repository as a whole is
considered.
Even if that job fails, the auto-comments/suggestions (made by
reviewdog) will still be made in the `format-suggestions-on-pr` workflow.